% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_download.R
\name{clean_exp}
\alias{clean_exp}
\title{Clean raw expression matrix}
\usage{
clean_exp(df)
}
\arguments{
\item{df}{Data frame. A data frame containing a \code{gene_id} column.}
}
\value{
A cleaned data frame with certain rows removed.
}
\description{
This function removes unwanted gene rows from a raw expression count or TPM matrix,
including pseudo-autosomal region genes and ambiguous counts.
}
