% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collectGarbage}
\alias{collectGarbage}
\title{Iterative garbage collection.}
\usage{
collectGarbage()
}
\value{
No return value.
}
\description{
Performs garbage collection until free memory idicators show no change.
same with \code{WGCNA::collectGarbage()}.
}
\examples{
collectGarbage()
}
