% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_integer}
\alias{get_integer}
\title{Extract Integer from XML Node}
\usage{
get_integer(doc, xpath)
}
\arguments{
\item{doc}{An XML document parsed by \code{xml2::read_xml()}.}

\item{xpath}{A character string representing the XPath query.}
}
\value{
An integer vector of node values. Returns \code{NA} if the node is not found.
}
\description{
A helper function to extract integer values from XML nodes using XPath.
}
\keyword{internal}
