% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_download.R
\name{get_manifest}
\alias{get_manifest}
\title{Extract and optionally write manifest from a GDC query result}
\usage{
get_manifest(query, manifest, download_dir = "GDCdata")
}
\arguments{
\item{query}{A query object returned by \code{TCGAbiolinks::GDCquery()}.}

\item{manifest}{(Optional) A path to save a manifest file. If missing,
the manifest will be generated and written to disk.}

\item{download_dir}{Directory to store downloaded files. Default is \code{"GDCdata"}.}
}
\value{
If \code{manifest} is missing, returns the file path to the saved manifest file.
Otherwise, returns the original manifest path.
}
\description{
This function extracts the manifest (including file ID, name, md5, size, and state)
from the \code{TCGAbiolinks::GDCquery()} result. If the \code{manifest} argument is not provided,
the manifest will be saved as a tab-delimited file in the working directory.
}
