% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_download.R
\name{get_tcga_mirna_rpm}
\alias{get_tcga_mirna_rpm}
\title{Extract RPM expression matrix from TCGA miRNA}
\usage{
get_tcga_mirna_rpm(project, manifest, download_dir = "GDCdata")
}
\arguments{
\item{project}{Character. TCGA project ID.}

\item{manifest}{A file path of the manifest file used for download.}

\item{download_dir}{Character. Directory for downloaded files. Default is \code{"GDCdata"}.}
}
\value{
A data frame of RPM values with gene annotations.
}
\description{
This function downloads and parses FPKM-normalized expression data from TCGA.
}
