% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_download.R
\name{get_tcga_mrna_count}
\alias{get_tcga_mrna_count}
\title{Extract raw expression counts from TCGA mRNA}
\usage{
get_tcga_mrna_count(project, manifest, download_dir = "GDCdata", clean = TRUE)
}
\arguments{
\item{project}{Character. TCGA project ID.}

\item{manifest}{A file path of the manifest file used for download.}

\item{download_dir}{Character. Directory for downloaded files. Default is \code{"GDCdata"}.}

\item{clean}{Logical. If \code{TRUE}, remove unwanted gene entries. Default is \code{TRUE}.}
}
\value{
A data frame of raw expression counts with gene annotations.
}
\description{
This function downloads and parses raw gene expression count data from TCGA.
}
