% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%<>\%}
\alias{\%<>\%}
\title{Pipe assignment operator}
\usage{
lhs \%<>\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using magrittr semantics.}
}
\value{
The result of assigning \code{rhs(lhs)} to \code{lhs}.
}
\description{
Re-export of the magrittr compound assignment pipe operator \code{\%<>\%}.
See \code{\link[magrittr:compound]{magrittr::\%<>\%}} for details.
}
