% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_download.R
\name{tcga_download}
\alias{tcga_download}
\title{Download TCGA data using TCGAbiolinks}
\usage{
tcga_download(
  project,
  data.category,
  data.type,
  data.format,
  workflow.type,
  query_rds,
  manifest,
  manifest_only = FALSE,
  download_dir = "GDCdata"
)
}
\arguments{
\item{project}{A character string specifying the TCGA project, e.g., \code{"TCGA-LGG"}.}

\item{data.category}{The data category, e.g., \code{"Clinical"}.}

\item{data.type}{The data type, e.g., \code{"Clinical Supplement"}.}

\item{data.format}{The file format, e.g., \code{"BCR XML"}.}

\item{workflow.type}{GDC workflow type.}

\item{query_rds}{save query to RDS.}

\item{manifest}{(Optional) A path to save a manifest file. If missing, it will be generated.}

\item{manifest_only}{Logical; if \code{TRUE}, only the manifest file will be generated without downloading data.}

\item{download_dir}{Directory to store downloaded files. Default is \code{"GDCdata"}.}
}
\value{
Returns the internal query object invisibly. Nothing is printed unless assigned to a variable.
}
\description{
This function queries and optionally downloads TCGA data from GDC using the TCGAbiolinks package.
It supports retrieving only the manifest file or both the manifest and data files.
}
