% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_download.R
\name{tcga_download_cnv}
\alias{tcga_download_cnv}
\title{Download TCGA copy number variation (CNV) data}
\usage{
tcga_download_cnv(
  project,
  manifest,
  download_dir = "GDCdata",
  manifest_only = FALSE
)

tcga_download_cnv(
  project,
  manifest,
  download_dir = "GDCdata",
  manifest_only = FALSE
)
}
\arguments{
\item{project}{TCGA project ID (e.g., "TCGA-CHOL").}

\item{manifest}{A file path of the manifest file used for download.}

\item{download_dir}{Directory where the downloaded files will be saved. Default is "GDCdata".}

\item{manifest_only}{Logical. If TRUE, only the manifest file is generated without downloading the data. Default is FALSE.}

\item{data.type}{The type of CNV data to download ("Copy Number Segment" or "Masked Copy Number Segment"). Default is "Copy Number Segment".}

\item{workflow.type}{Workflow type used to generate the CNV data ("DNAcopy" or "GATK4 CNV"). Default is "DNAcopy".}
}
\value{
Returns invisibly the query object created by \code{tcga_download()}.

Returns invisibly the query object created by \code{tcga_download()}.
}
\description{
This function downloads CNV data from TCGA using a manifest file.

This function downloads protein expression quantification data from TCGA using a manifest file.
}
