% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_download.R
\name{tcga_download_methy_beta}
\alias{tcga_download_methy_beta}
\title{Download TCGA DNA methylation beta values}
\usage{
tcga_download_methy_beta(
  project,
  manifest,
  download_dir = "GDCdata",
  platform = "Illumina Human Methylation 450",
  manifest_only = FALSE
)
}
\arguments{
\item{project}{TCGA project ID (e.g., "TCGA-CHOL").}

\item{manifest}{A file path of the manifest file used for download.}

\item{download_dir}{Directory where the downloaded data will be saved. Default is "GDCdata".}

\item{platform}{Platform name for DNA methylation data (e.g., "Illumina Human Methylation 450" or "Illumina Human Methylation 27").}

\item{manifest_only}{Logical. If TRUE, only the manifest file is generated without actual download. Default is FALSE.}
}
\value{
Returns invisibly the query object generated by \code{tcga_download()}.
}
\description{
This function downloads DNA methylation beta value data from TCGA using a manifest file.
}
\examples{
\dontrun{

query <- tcga_download_methy_beta(project = "TCGA-CHOL", platform = "Illumina Human Methylation 450")
beta <- TCGAbiolinks::GDCprepare(query, directory = "GDCdata", summarizedExperiment = FALSE)

}

}
