% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_download.R
\name{tcga_download_mirna_seq_exp}
\alias{tcga_download_mirna_seq_exp}
\title{Download TCGA miRNA-seq expression data}
\usage{
tcga_download_mirna_seq_exp(
  project,
  manifest,
  download_dir = "GDCdata",
  manifest_only = FALSE
)
}
\arguments{
\item{project}{TCGA project ID (e.g., "TCGA-BRCA").}

\item{manifest}{A file path of the manifest file used for download.}

\item{download_dir}{Directory where the downloaded data will be saved. Default is "GDCdata".}

\item{manifest_only}{Logical. If TRUE, only the manifest file is generated without actual download. Default is FALSE.}
}
\value{
Returns invisibly the query object generated by \code{tcga_download()}.
}
\description{
This function downloads miRNA expression quantification data from TCGA using a manifest file.
}
