% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_download.R
\name{tcga_download_snp}
\alias{tcga_download_snp}
\title{Download TCGA somatic mutation (SNP) data}
\usage{
tcga_download_snp(
  project,
  manifest,
  download_dir = "GDCdata",
  manifest_only = FALSE
)
}
\arguments{
\item{project}{TCGA project ID (e.g., "TCGA-CHOL").}

\item{manifest}{A file path of the manifest file used for download.}

\item{download_dir}{Directory where the downloaded files will be stored. Default is "GDCdata".}

\item{manifest_only}{Logical. If TRUE, only generates the manifest file without downloading data. Default is FALSE.}
}
\value{
Returns invisibly the query object created by \code{tcga_download()}.
}
\description{
This function downloads masked somatic mutation data (SNP) from TCGA using a manifest file.
Only open-access MAF files are supported.
}
