% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_clinical_xml.R
\name{tcga_xml2list}
\alias{tcga_xml2list}
\title{Parse Multiple TCGA XML Files into List}
\usage{
tcga_xml2list(xfiles, verbose = FALSE)
}
\arguments{
\item{xfiles}{A character vector of file paths to TCGA XML clinical files.}

\item{verbose}{print information of file path, default is FALSE.}
}
\value{
A list where each element is the output of \code{tcga_xml_parse()}.
}
\description{
Applies \code{tcga_xml_parse()} to a vector of TCGA XML clinical files.
}
\seealso{
\code{\link{tcga_xml_parse}}, \code{\link{tcga_xml2df}}, \code{\link{tcga_list2df}}
}
